Refactor suse manager cleaner
