+++ jenkins_pipelines/scripts/SUSEManager_cleaner/README.md
