+++ jenkins_pipelines/scripts/SUSEManager_cleaner/requirements.txt
