+++ jenkins_pipelines/scripts/SUSEManager_cleaner/suse_manager_cleaner_program/__init__.py
