+++ jenkins_pipelines/scripts/SUSEManager_cleaner/suse_manager_cleaner_program/suse_manager_ssh.py
