+++ jenkins_pipelines/scripts/SUSEManager_cleaner/tests/__init__.py
